Allow Bluetooth functionality in work profiles
This dirty change modifies every reference throughout the Bluetooth
module that checks if the caller is the system user or the active
user, now permitting calls from a managed profile as well.

This is handled in the functions 'checkCallerIsPermittedUser' and
'callerIsPermittedUser', which currently behave the same as the
'checkCallerIsSystemOrActiveOrManagedUser' function, but in the
future could be altered to take other factors into account.

Change-Id: Ie16714c300293ffe22fb8b0502ccf6f02bbeffcf